Move nodemon to devDependencies
Installing nodemon in the Dockerfile was a hack to get things running. The
dependency has now been moved to package.json and the necessary environment
variables have been set.

I'm still not happy with how dependencies are loaded as the COPY command feels
dirty and will require a rebuild on updated dependencies. But copying the json
file seems to be the standard way of loading node dependencies at the moment.

To rebuild the image run
    docker-compose build --no-cache